Changes to buildroot packages: libgtk2 and cairo
Changes to libgtk2/config.in are to remove the selection of X lib.
Changes to libgtk2/libgth2.mk are to remove xlib from the build and add direct
fb and also to remove the code that deletes gtk-demo from the install target.
Changes to cairo/cario.mk are to remove the requirement of having the package
directfb_svg for selecting to use directfb.